r/aws_client_vpn_endpoint: Add 'testAccClientVPNEndpoint_simpleAttributesUpdate'.
Acceptance test output:

% make testacc TESTS=TestAccEC2ClientVPNEndpoint_serial/Endpoint_simpleAttributesUpdate PKG=ec2
==> Checking that code complies with gofmt requirements...
TF_ACC=1 go test ./internal/service/ec2/... -v -count 1 -parallel 20 -run='TestAccEC2ClientVPNEndpoint_serial/Endpoint_simpleAttributesUpdate'  -timeout 180m
=== RUN   TestAccEC2ClientVPNEndpoint_serial
=== PAUSE TestAccEC2ClientVPNEndpoint_serial
=== CONT  TestAccEC2ClientVPNEndpoint_serial
=== RUN   TestAccEC2ClientVPNEndpoint_serial/Endpoint_simpleAttributesUpdate
=== PAUSE TestAccEC2ClientVPNEndpoint_serial/Endpoint_simpleAttributesUpdate
=== CONT  TestAccEC2ClientVPNEndpoint_serial/Endpoint_simpleAttributesUpdate
--- PASS: TestAccEC2ClientVPNEndpoint_serial (0.86s)
    --- PASS: TestAccEC2ClientVPNEndpoint_serial/Endpoint_simpleAttributesUpdate (34.32s)
PASS
ok      github.com/hashicorp/terraform-provider-aws/internal/service/ec2        38.419s

+++ internal/service/ec2/client_vpn_endpoint.go
@@ -300,7 +300,7 @@ func resourceClientVPNEndpointUpdate(d *schema.ResourceData, meta interface{}) e
 		}
 
 		if d.HasChange("session_timeout_hours") {
-			input.SessionTimeoutHours = aws.Int64(d.Get("session_timeout_hours").(int64))
+			input.SessionTimeoutHours = aws.Int64(int64(d.Get("session_timeout_hours").(int)))
 		}
 
 		if d.HasChange("server_certificate_arn") {
